articlesdatabase/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-filter==23.5
pandas==2.2.1
djangorestframework.authtoken
djangorestframework-simplejwt>=5.2.0
# django-cors-headers==4.3.1
# django-debug-toolbar==4.3.0